requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.7
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
idna==3.7
importlib_resources==6.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.1
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
setuptools==71.0.3
simplejson==3.19.2
sqlparse==0.5.1
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
